Change: split main.go
